Reuse sort functions for common, popular, and alphabetical lists
